+++ modules/kubernetes/eks-core/modules.tf
@@ -23,7 +23,6 @@ locals {
     "external-dns",
     "falco,flux-system",
     "gatekeeper-system",
-    "goldpinger",
     "ingress-healthz",
     "ingress-nginx",
     "kube-node-lease",
@@ -309,7 +308,6 @@ module "prometheus" {
   falco_enabled                          = var.falco_enabled
   opa_gatekeeper_enabled                 = var.opa_gatekeeper_enabled
   linkerd_enabled                        = var.linkerd_enabled
-  goldpinger_enabled                     = var.goldpinger_enabled
   flux_enabled                           = var.fluxcd_v2_enabled
   csi_secrets_store_provider_aws_enabled = var.csi_secrets_store_provider_aws_enabled
   azad_kube_proxy_enabled                = var.azad_kube_proxy_enabled
@@ -382,20 +380,6 @@ module "xenit" {
   }
   thanos_receiver_fqdn = var.xenit_config.thanos_receiver
   loki_api_fqdn        = var.xenit_config.loki_api
-}
-
-module "goldpinger" {
-  depends_on = [module.opa_gatekeeper, module.linkerd]
-
-  for_each = {
-    for s in ["goldpinger"] :
-    s => s
-    if var.goldpinger_enabled
-  }
-
-  source = "../../kubernetes/goldpinger"
-
-  linkerd_enabled = var.linkerd_enabled
 }
 
 module "new_relic" {