+++ modules/kubernetes/eks-core/variables.tf
@@ -315,12 +315,6 @@ variable "linkerd_enabled" {
   default     = false
 }
 
-variable "goldpinger_enabled" {
-  description = "Should goldpinger be enabled"
-  type        = bool
-  default     = true
-}
-
 variable "csi_secrets_store_provider_aws_enabled" {
   description = "Should csi-secrets-store-provider-aws be enabled"
   type        = bool